+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Detect suspected duplicate GML files (same size, different name)
+- Support multiple GML files import with glob patterns (e.g. `rcn_*.gml`)
 
 ## [0.1.0] - 2026-02-21